+++ NEWS.md
@@ -4,11 +4,11 @@
 
 ## Bug fixes
 
-* Revision of one-sided p-value computation for Log-Rank test is median TTE is NA
+* Revision of one-sided p-value computation for Log-Rank test
 
 * Revision of the call for Statistic in the core function (not visible)
 
-* Revision of the function to calculate the Hazard Ratio Statistic (HazardRatioStat method). This calculation is now based on the log-rank test ((O2/E2)/(O1/E1) where O and E are Observed and Expected event in group 2 and group 1.
+* Revision of the function to calculate the Hazard Ratio Statistic (HazardRatioStat method). This calculation is now based on the log-rank test ((O2/E2)/(O1/E1) where O and E are Observed and Expected event in sample 2 and sample 1.
 
 * Revision of the function to generate the Test, Statistic, Design and result tables in the report